Updated all test cases and fixed all errors and warnings
