+++ src/mockData/articleListData.json
@@ -73,7 +73,7 @@
       "author": "Benzinga Neuro"
     }
   ],
-  "articlesToDisplay": [
+  "activePageArticles": [
     {
       "title": "Huawei Chips Away At Apple, Nvidia: Revenue Nears $100B As China Flexes Tech Muscle Despite US Sanctions - NVIDIA  ( NASDAQ:NVDA )",
       "url": "https://www.benzinga.com/news/23/12/36421258/huawei-chips-away-at-apple-nvidia-revenue-nears-100b-as-china-flexes-tech-muscle-despite-us-sanction",
