+++ src/tests/components/FilterList.test.tsx
@@ -1,33 +1,24 @@
-import { render, screen, within } from "@testing-library/react";
+import { render, screen } from "@testing-library/react";
+import { Provider } from "react-redux";
+import store from "../../store";
 import FilterList from "../../components/FilterList";
-import filterListData from "../../mockData/filterListData.json";
+
+// Mock the axios module
+jest.mock("axios", () => ({
+  __esModule: true,
+  default: {
+    get: jest.fn(),
+  },
+}));
 
 describe("Filter List component test suite", () => {
-  it("renders category filter as string with correct value", () => {
-    render(<FilterList />);
-    const categoryFilter = screen.getByTestId("category-filter");
-    const categoryFilterItems =
-      within(categoryFilter).getAllByTestId("filter-item");
-    const categoryLabel = within(categoryFilterItems[0]).getByTestId(
-      "filter-label"
+  it("renders filter list wrapper correctly", () => {
+    render(
+      <Provider store={store}>
+        <FilterList />
+      </Provider>
     );
-    expect(categoryFilterItems).toHaveLength(filterListData.categories.length);
-    expect(categoryLabel).toBeInTheDocument();
-    expect(typeof categoryLabel.textContent).toBe("string");
-    expect(categoryLabel.textContent).toBe(filterListData.categories[0]);
-  });
-
-  it("renders author filter as string with correct value", () => {
-    render(<FilterList />);
-    const authorFilter = screen.getByTestId("author-filter");
-    const authorFilterItems =
-      within(authorFilter).getAllByTestId("filter-item");
-    const authorLabel = within(authorFilterItems[0]).getByTestId(
-      "filter-label"
-    );
-    expect(authorFilterItems).toHaveLength(filterListData.authors.length);
-    expect(authorLabel).toBeInTheDocument();
-    expect(typeof authorLabel.textContent).toBe("string");
-    expect(authorLabel.textContent).toBe(filterListData.authors[0]);
+    const filterListWrapper = screen.getByTestId("filter-list-wrapper");
+    expect(filterListWrapper).toBeInTheDocument();
   });
 });
